% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pal_gradient}
\alias{pal_gradient}
\title{Return interpolated color gradient for a continuous poisson colour palette}
\usage{
pal_gradient(n = 256, palette, reverse = FALSE)
}
\arguments{
\item{n}{Number of steps in gradient}

\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colourRampPalette()}
}
\description{
Return interpolated color gradient for a continuous poisson colour palette
}
