% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pois_colours}
\alias{pois_colours}
\title{Poisson Colours}
\format{An object of class \code{character} of length 29.}
\usage{
pois_colours
}
\description{
@format A named character vector of Poisson Colours as hex codes
 @examples 
 .pois_colours
}
\keyword{datasets}
