% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pois_cols}
\alias{pois_cols}
\title{poisson colours}
\usage{
pois_cols(colours = NULL)
}
\arguments{
\item{colours}{Character names of Poisson colours}
}
\value{
A named character vector of hex colours
}
\description{
poisson colours
}
\examples{
pois_cols()
pois_cols(c("red", "yellow"))
}
