% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pois_pal}
\alias{pois_pal}
\title{Return function to interpolate a poisson colour palette}
\usage{
pois_pal(palette = "legacy1", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colourRampPalette()}
}
\description{
Return function to interpolate a poisson colour palette
}
