% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pois_pal_custom}
\alias{pois_pal_custom}
\title{Return function to subset a user provided colour palette}
\usage{
pois_pal_custom(palette, order = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character vector of hex colours.}

\item{order}{A numeric or character vector indicating the order of colours in the palette. Can be a subset.}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\value{
A function that takes the number of colours required and returns a
character vector of that many hex colours from the palette.
}
\description{
Return function to subset a user provided colour palette
}
\examples{
pois_pal_custom(c("#1E6091", "#D9ED92", "#99D98C"))(3)
}
