% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pois_pal_disc}
\alias{pois_pal_disc}
\title{Return function to subset poisson colour palette}
\usage{
pois_pal_disc(palette = "discrete", order = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette. See \code{pois_pal()} for palette options.}

\item{order}{A numeric or character vector indicating the order of colours in the palette. Can be a subset.}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\value{
A function that takes the number of colours required and returns a
character vector of that many hex colours from the palette.
}
\description{
Return function to subset poisson colour palette
}
\examples{
pois_pal_disc("discrete")(3)
}
