% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pois_palettes}
\alias{pois_palettes}
\title{Poisson Palettes}
\format{An object of class \code{list} of length 6.}
\usage{
pois_palettes
}
\description{
@format A list of character vectors of Poisson colour palettes
 @examples 
 .pois_palettes
}
\keyword{datasets}
