% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_disc_pois}
\alias{scale_fill_disc_pois}
\title{Fill scale constructor for poisson colours}
\usage{
scale_fill_disc_pois(palette = "colours", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Fill scale constructor for poisson colours
}
