% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_disc_poisson}
\alias{scale_fill_disc_poisson}
\title{discrete fill scale constructor for poisson colours}
\usage{
scale_fill_disc_poisson(
  ...,
  palette = getOption("poispalette.colours", "discrete"),
  order = NULL,
  reverse = FALSE
)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale()}

\item{palette}{Character name of palette in pois_palettes}

\item{order}{A numeric or character vector indicating the order of colours in the palette. Can be a subset.}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
discrete fill scale constructor for poisson colours
}
