% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_pois}
\alias{scale_fill_pois}
\title{Fill scale constructor for poisson colors}
\usage{
scale_fill_pois(palette = "legacy1", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for poisson colors
}
