% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pois_colours}
\alias{pois_colours}
\title{Poisson Colours}
\format{A named character vector of Poisson Colours as hex codes}
\usage{
pois_colours
}
\description{
Poisson Colours
}
\examples{
pois_colours
}
\keyword{datasets}
