% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pois_cols}
\alias{pois_cols}
\title{poisson colours}
\usage{
pois_cols(cols = NULL)
}
\arguments{
\item{cols}{Character names of Poisson colours}
}
\value{
a named vector of hex colours
}
\description{
poisson colours
}
\examples{
pois_cols()
}
