% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{pois_pal}
\alias{pois_pal}
\title{poisson palettes}
\usage{
pois_pal(palette = NULL)
}
\arguments{
\item{palette}{Character names of Poisson palettes. One of 'discrete', 'hot', 'cool'.}
}
\value{
A named vector of hex colours
}
\description{
poisson palettes
}
\examples{
pois_pal()
}
