% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{pois_pal_custom}
\alias{pois_pal_custom}
\title{Return function to subset a user provided colour palette}
\usage{
pois_pal_custom(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colourRampPalette()}
}
\description{
Return function to subset a user provided colour palette
}
