% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{pois_pal_custom}
\alias{pois_pal_custom}
\title{Return function to subset a user provided colour palette}
\usage{
pois_pal_custom(palette, order = NULL, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{order}{A numeric or character vector indicating the order of colours in the palette. Can be a subset.}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colourRampPalette()}
}
\description{
Return function to subset a user provided colour palette
}
