% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{pois_pal_disc}
\alias{pois_pal_disc}
\title{Return function to subset poisson colour palette}
\usage{
pois_pal_disc(palette = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. See \code{pos_pal()} for palette options.}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colourRampPalette()}
}
\description{
Return function to subset poisson colour palette
}
