% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{pois_pal_grad}
\alias{pois_pal_grad}
\title{Return interpolated color gradient for a continuous poisson colour palette}
\usage{
pois_pal_grad(n_steps = 256, palette, reverse = FALSE)
}
\arguments{
\item{n_steps}{Number of steps in gradient}

\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
Return interpolated color gradient for a continuous poisson colour palette
}
