% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pois_palettes}
\alias{pois_palettes}
\title{Poisson Palettes}
\format{A list of named character vectors of Poisson colour palettes}
\usage{
pois_palettes
}
\description{
Poisson Palettes
}
\examples{
pois_palettes
}
\keyword{datasets}
