% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poispalette-package.R
\docType{package}
\name{poispalette-package}
\alias{poispalette}
\alias{poispalette-package}
\title{poispalette: Poisson Palettes}
\description{
An R package for Poisson Consulting colour palettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/poispalette}
  \item Report bugs at \url{https://github.com/poissonconsulting/poispalette/issues}
}

}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evan@poissonconsulting.ca} (\href{https://orcid.org/-3188}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
