% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_disc_poisson}
\alias{scale_colour_disc_poisson}
\title{discrete colour scale constructor for poisson colours}
\usage{
scale_colour_disc_poisson(
  palette = getOption("poispalette.colours", "colours"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in pois_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
discrete colour scale constructor for poisson colours
}
